# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Phase Two Cluster Management API
  description: |-
    Automate management of your Phase Two hosted Keycloak clusters: dedicated clusters, deployments (realms), custom domains, extensions, environment variables, IP rules, organizations, and billing.

    ## Authentication

    All requests require an OAuth2 access token obtained via the **client credentials grant** against your Phase Two organization's OIDC token endpoint (`https://<your-console-host>/realms/self/protocol/openid-connect/token`, where `<your-console-host>` is the same host you use to log into the Phase Two console). Create a client (an "API secret") for your organization via the `org.apiSecret.create` operation below, then request a token with that client's ID and secret using `grant_type=client_credentials`. Send the resulting token as `Authorization: Bearer <token>` on every call.
  version: 1.0.0
servers:
- url: "https://{environment}.phasetwo.io/auth/realms/self/v2"
  description: Phase Two hosted console API.
  variables:
    environment:
      default: app
      description: "`app` for production, `app-staging` for staging. This is the same\
        \ host you use to log into the Phase Two console."
      enum:
      - app
      - app-staging
security:
- oidcClientCredentials: []
...
paths:
...
  /clusters/{id}/extensions:
    get:
      tags:
      - extensions
      summary: List extensions for a cluster.
      description: "Returns the custom extensions (providers, themes) configured for\
        \ this cluster, optionally filtered by type."
      operationId: cluster.extension.list
      parameters:
      - name: type
        in: query
        description: Only return extensions of this type.
        schema:
          $ref: "#/components/schemas/ResourceType"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Extension"
    post:
      tags:
      - extensions
      summary: Create an extension for a cluster.
      description: Registers a new extension slot (provider or theme) for this cluster.
        This only creates the extension record; upload its jar via the extension's
        version sub-resource before it takes effect.
      operationId: cluster.extension.create
      requestBody:
        description: The extension to create.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtensionRequest"
        required: true
      responses:
        "201":
          description: "Extension created. The response has no body; `Location` holds\
            \ the new extension's URL, whose last path segment is its ID."
          headers:
            Location:
              description: URL of the created extension.
              style: simple
              schema:
                type: string
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
  /clusters/{id}/extensions/cluster-update:
    post:
      tags:
      - extensions
      summary: Restart a cluster to apply extension changes.
      description: Reconciles this cluster's configured extensions onto its running
        Keycloak deployment and restarts it to pick up the changes. Fails with a 409
        if a restart is already in progress for this cluster.
      operationId: cluster.extension.reconcile
      responses:
        "202":
          description: Restart accepted.
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
...
  /clusters/{id}/extensions/{extensionId}:
    get:
      tags:
      - extensions
      summary: Get an extension by ID.
      operationId: cluster.extension.detail
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Extension"
    put:
      tags:
      - extensions
      summary: Enable or disable an extension.
      description: Toggles whether an extension is active. Disabling does not delete
        its versions.
      operationId: cluster.extension.update
      requestBody:
        description: The desired enabled state.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleExtensionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Extension"
    delete:
      tags:
      - extensions
      summary: Remove an extension by ID.
      description: Deletes an extension and all of its versions.
      operationId: cluster.extension.delete
      responses:
        "204":
          description: Extension removed.
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    - name: extensionId
      in: path
      required: true
      schema:
        type: string
...
components:
  schemas:
...
    Extension:
      required:
      - id
      - name
      - enabled
      - resourceType
      type: object
      properties:
        id:
          description: Unique identifier of the extension.
          type: string
        name:
          description: "Extension name, unique within the cluster."
          type: string
        enabled:
          description: Whether the extension is active on the cluster.
          type: boolean
        created_at:
          format: int64
          description: "When the extension was created, as epoch milliseconds."
          type: integer
        versions:
          description: Uploaded versions of this extension. A version-independent
            extension has at most one.
          type: array
          items:
            $ref: "#/components/schemas/ExtensionVersion"
        resourceType:
          description: What kind of resource this is — determines whether versions
            are Keycloak-version-dependent.
          type: string
          allOf:
          - $ref: "#/components/schemas/ResourceType"
    ExtensionRequest:
      required:
      - name
      - resourceType
      type: object
      properties:
        name:
          description: "Name for the extension. Max 63 characters, lowercase letters\
            \ only, no special characters."
          type: string
        resourceType:
          description: What kind of resource this is — determines whether versions
            are Keycloak-version-dependent.
          type: string
          allOf:
          - $ref: "#/components/schemas/ResourceType"
    ExtensionVersion:
      required:
      - id
      - valid
      - scan_state
      type: object
      properties:
        id:
          description: Unique identifier of the extension version.
          type: string
        valid:
          description: "Whether this version is approved for deployment. This is the\
            \ human approval decision, not the scanner's verdict — see `scan_state`."
          type: boolean
        label:
          description: Label recorded when the jar was uploaded.
          type: string
        resourceKey:
          description: S3 object key of the uploaded jar. Absent until an upload is
            confirmed.
          type: string
        keycloakMajorVersion:
          format: int32
          description: Keycloak major version this version targets. Absent for version-independent
            extensions.
          type: integer
        scan_state:
          description: "Result of the security scan: `UNSCANNED`, `SCANNING`, `PASS`,\
            \ `WARN`, `BLOCK` or `ERROR`. Never absent."
          type: string
        risk_score:
          format: int32
          description: Risk score from 0-100 reported by the scan. Absent until a
            scan has completed.
          type: integer
        report_location:
          description: "Location of the full scan report, when one exists."
          type: string
        scanned_at:
          format: int64
          description: "When the version was last scanned, as epoch milliseconds."
          type: integer
        scanner_version:
          description: Version of the scanner that produced the result.
          type: string
        override_by:
          description: ID of the user who approved this version despite the scan not
            passing. Cleared when the approval is withdrawn.
          type: string
        override_reason:
          description: Justification recorded with that approval.
          type: string
        created_at:
          format: int64
          description: "When the version was created, as epoch milliseconds."
          type: integer
...
    ResourceType:
      description: Kind of custom resource an extension holds. `THEME` and `EXTENSION`
        are tied to a Keycloak major version and use the versioned upload flow; `PASSWORD_BLACKLIST`
        and `WELL_KNOWN` are version-independent and use the standalone upload flow.
      enum:
      - THEME
      - EXTENSION
      - PASSWORD_BLACKLIST
      - WELL_KNOWN
      type: string
...
    ToggleExtensionRequest:
      required:
      - enabled
      type: object
      properties:
        enabled:
          description: Whether the extension should be active on the cluster.
          type: boolean
...
  securitySchemes:
    oidcClientCredentials:
      type: oauth2
      description: "OIDC client credentials grant. Obtain a client ID/secret pair\
        \ via the `org.apiSecret.create` operation, then exchange them for an access\
        \ token at your realm's `/protocol/openid-connect/token` endpoint."
      flows:
        clientCredentials:
          tokenUrl: https://<your-console-host>/realms/self/protocol/openid-connect/token